requirements.txt
PyQt5~=5.15.1
psycopg2-binary~=2.8.6